# Learn (Data Types) Lesson

+++ src/_6_variables/Variables4.cpp
@@ -2,8 +2,8 @@ using namespace std;
 #include <iostream>
 
 int main() {
-    const int INCREASE_YEARS = 5;
-    const int AGE = 22;
-    const int AGE_AFTER_INCREASE = AGE + INCREASE_YEARS;
+    const short INCREASE_YEARS = 5;
+    const short AGE = 22;
+    const short AGE_AFTER_INCREASE = AGE + INCREASE_YEARS;
     cout << "After " << INCREASE_YEARS << " years you will be " << AGE_AFTER_INCREASE;
 }
